+++ src/game_popups.cpp
@@ -121,7 +121,7 @@ void game_popups_display(Gamep g, Levelsp v, Levelp l)
         // Rise up
         //
         const auto tile_height   = br.y - tl.y;
-        auto       ascend_height = (int) (pct * (float) tile_height * POPUP_DURATION_TILE_HEIGHT);
+        auto       ascend_height = (int) (pct * (float) tile_height * (float) POPUP_DURATION_TILE_HEIGHT);
         tl.y -= ascend_height;
         br.y -= ascend_height;
