+++ src/level_light.cpp
@@ -519,7 +519,7 @@ void Raycast::raycast_do(Gamep g, Levelsp v, Levelp l)
       // Once we hit an obstacle to vision, how far do we allow the ray of light to penetrate
       //
       auto obs_to_vision_start_distance       = ray_pixel->distance;
-      auto obs_to_vision_penetration_distance = (obs_to_vision_start_distance + TILE_WIDTH) - 2;
+      auto obs_to_vision_penetration_distance = (obs_to_vision_start_distance + (float) TILE_WIDTH) - 2;
 
       //
       // Keep track of the type of object we hit
