+++ src/tests/test_alloc_free_wall.cpp
@@ -47,7 +47,7 @@
       // No need to test all slots for level select. Additionally, we've filled
       // all the other levels, so not a lot of space to alloc IDs from.
       //
-      auto slot_depth_to_test = MAP_SLOTS;
+      auto slot_depth_to_test = (int) MAP_SLOTS;
       if (level_num == LEVEL_SELECT_ID) {
         //
         // There's just no space in other levels if we fully allocate them!
